Revert "Merge pull request #18 from grafana/mmandrus/client-reconnect"
This reverts the change to attempt to reconnect within this client.
The motivation for this is to avoid spawning a new goroutine when
there are connection errors. Similar functionality exists in the
wrapper of this client in dskit which automatically re-resolves
servers every 30 seconds and updates the selector's list of servers.

This reverts commit fdaf6a95408eab4f992f047f7f037b32d9927f3b, reversing
changes made to e553a78ea238794b71fa2ced9ba64b157465fb70.

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>